websocket/realtime-transcriptions/requirements.txt
Flask==1.0.2
Flask-Sockets==0.2.1
google-cloud-speech==0.36.3
gunicorn==19.9.0
